feat(resolver): add support for minitest
Wouldn't it be great if this extension supported Minitest, too? You bet it would! Here's how:

`isSpec()` is easy enough to get working by checking for `indexOf('_test.rb')`.

`specToCode()` is a little more tricky, but we can check for a regex match to get what we need.

`codeToSpec()` requires a little more work, but we can check if the `spec` directory exists to get what we need.

**NOTE**: I tried to pull in Sinon.js to stub `_specDirExists()`, but `npm run comile` kept barking about "cannot find module sinon". After a fair bit of googling, I'm not sure how to solve this. Any ideas are appreciated. :smil: